// templates/smm-store/src/lib/types.ts
export interface ExternalSource {
  id: number;
  name: string;
  icon: string;
  type: string;
  url: string;
  status: string;
  statusColor: string;
  lastSync: string;
  products: number;
  balance: string;
  profitPercentage?: number;
  profitAmount?: number | null;
  connectionError?: string | null;
  syncOnly?: boolean;
  connectionStatus?: string;
}